requirements.txt
torch
torchvision
transformers
PyYAML
pillow
tqdm
clip-by-openai
